fix(migrations): wiki_pages.version must be BIGINT not INTEGER (Postgres i64)
CI run 25007588682 failed wiki_store_search_against_postgres_tsvector
(7 of 8 tests passed; 1 remaining):

  Sqlx(ColumnDecode { index: "\"version\"",
    source: "mismatched types; Rust type `i64` (as SQL type `INT8`)
            is not compatible with SQL type `INT4`" })

The Rust `WikiPage::version` and `WikiPageVersion::version` fields are
`i64`. Postgres `INTEGER` is 32-bit (`INT4`), and sqlx refuses to decode
that into `i64` even when the value fits. SQLite's INTEGER is
dynamically-sized so the SQLite arm decodes cleanly without per-backend
casts.

Fix: declare both `wiki_pages.version` and `wiki_page_versions.version`
as BIGINT in the Postgres migration. Same precedent as
`audit_events.seq` (already BIGINT) and `task_number_seq.next_number`
(already BIGINT).

Other i64 / INTEGER mismatches checked and OK:
  - wiki_pages.archived (Rust `bool` via per-backend `i32` read at
    src/wiki/store.rs:1129; Postgres path takes the i32 branch)
  - projects.sort_order (Rust `i64` via per-backend `i32 as i64` cast at
    src/projects/store.rs:973; INTEGER → i32 → i64 is a valid widening)
  - task_number_seq.id (singleton CHECK (id = 1); never deserialized)
  - task_number_seq.next_number (already BIGINT)

cargo check --test instance_postgres ✓ green.

+++ migrations/postgres/global/20260407120000_wiki.sql
@@ -25,7 +25,11 @@ CREATE TABLE IF NOT EXISTS wiki_pages (
     related TEXT NOT NULL DEFAULT '[]',
     created_by TEXT NOT NULL,
     updated_by TEXT NOT NULL,
-    version INTEGER NOT NULL DEFAULT 1,
+    -- BIGINT (vs SQLite's dynamic INTEGER) so sqlx's FromRow derive on
+    -- WikiPage::version maps cleanly without per-backend i32→i64 casts.
+    -- The Rust struct uses `i64` and Postgres `INTEGER` is 32-bit (INT4)
+    -- which sqlx refuses to decode into i64.
+    version BIGINT NOT NULL DEFAULT 1,
     archived INTEGER NOT NULL DEFAULT 0,
     created_at TEXT NOT NULL DEFAULT to_char(now() AT TIME ZONE 'UTC', 'YYYY-MM-DD"T"HH24:MI:SS.MS"Z"'),
     updated_at TEXT NOT NULL DEFAULT to_char(now() AT TIME ZONE 'UTC', 'YYYY-MM-DD"T"HH24:MI:SS.MS"Z"'),
@@ -42,7 +46,8 @@ CREATE TABLE IF NOT EXISTS wiki_pages (
 CREATE TABLE IF NOT EXISTS wiki_page_versions (
     id TEXT PRIMARY KEY DEFAULT replace(gen_random_uuid()::text, '-', ''),
     page_id TEXT NOT NULL REFERENCES wiki_pages (id),
-    version INTEGER NOT NULL,
+    -- BIGINT for parity with wiki_pages.version above; sqlx FromRow needs i64.
+    version BIGINT NOT NULL,
     content TEXT NOT NULL,
     edit_summary TEXT,
     author_type TEXT NOT NULL CHECK (
